+++ mobile/src/types/IssueCategoryArray.ts
@@ -1,4 +1,4 @@
-//mobile/src/types/IssueCategoryArray/ts
+//mobile/src/types/IssueCategoryArray.ts
 import { IssueCategory } from '@civickit/shared'
 
 export const IssueCategoryArray = [
